Address PR feedback
Confirmation Modals now show on top of other confirmation modals with same priority

+++ javascripts/core/app/modal.js
@@ -62,36 +62,36 @@ import PelleEffectsModal from "@/components/modals/PelleEffectsModal";
 
 
 export class Modal {
-  constructor(component, priority = 0, bare = false) {
+  constructor(component, priorityStatement = {}, bare = false) {
     this._component = component;
     this._bare = bare;
     this._modalConfig = {};
-    this._priority = priority;
+    this._priority = priorityStatement.priority ?? 0;
+    this._showFirst = priorityStatement.showFirst ?? false;
   }
 
   show(modalConfig) {
     if (!GameUI.initialized) return;
     this._props = Object.assign({}, modalConfig || {});
     const modalQueue = ui.view.modal.queue;
-    // New modals go to the back of the queue (shown last).
-    if (!modalQueue.includes(this)) {
-      if (!modalQueue.length || this._priority <= modalQueue.last()._priority) {
-        modalQueue.push(this);
-      } else {
-        for (let i = 0; i < modalQueue.length; i++) {
-          if (modalQueue[i]._priority < this._priority) {
-            modalQueue.splice(i, 0, this);
-            break;
-          }
+    if (modalQueue.includes(this)) return;
+
+    // Modals are first inserted into the front of the queue, then make their way back
+    // until they see the next Modal in the queue has a lower priority
+    // If the modal is shown first then it makes its way back it will insert itself if the next Modal
+    // has the same priority too
+    if (!modalQueue.length || (this._priority <= modalQueue.last()._priority && !this._showFirst)) {
+      modalQueue.push(this);
+    } else {
+      for (let i = 0; i < modalQueue.length; i++) {
+        if ((this._showFirst && (modalQueue[i]._priority <= this._priority)) ||
+          modalQueue[i]._priority < this._priority) {
+          modalQueue.splice(i, 0, this);
+          break;
         }
       }
-      ui.view.modal.current = modalQueue[0];
     }
-  }
-
-  setPriority(x) {
-    this._priority = x;
-    return this;
+    ui.view.modal.current = modalQueue[0];
   }
 
   get isOpen() {
@@ -142,6 +142,9 @@ class ChallengeConfirmationModal extends Modal {
   }
 }
 
+const PriorityPresets = {
+  confirm(priority) { return { priority, showFirst: true }; }
+};
 // If a new modal which can be shown in the same queue multiple times needs to be added
 // Additional code needs to be written to account for that
 
@@ -149,20 +152,20 @@ Modal.startEternityChallenge = new ChallengeConfirmationModal(EternityChallengeS
 Modal.startInfinityChallenge = new ChallengeConfirmationModal(InfinityChallengeStartModal);
 Modal.startNormalChallenge = new ChallengeConfirmationModal(NormalChallengeStartModal);
 
-Modal.dimensionBoost = new Modal(DimensionBoostModal, 1);
-Modal.antimatterGalaxy = new Modal(AntimatterGalaxyModal, 1);
-Modal.bigCrunch = new Modal(BigCrunchModal, 1);
-Modal.replicantiGalaxy = new Modal(ReplicantiGalaxyModal, 1);
-Modal.eternity = new Modal(EternityModal, 1);
-Modal.enterDilation = new Modal(EnterDilationModal, 1);
-Modal.reality = new Modal(RealityModal, 1);
-Modal.resetReality = new Modal(ResetRealityModal, 1);
-Modal.exitCelestialReality = new Modal(ExitCelestialModal, 1);
-Modal.celestials = new Modal(EnterCelestialsModal, 1);
-Modal.hardReset = new Modal(HardResetModal, 1);
+Modal.dimensionBoost = new Modal(DimensionBoostModal, PriorityPresets.confirm(1));
+Modal.antimatterGalaxy = new Modal(AntimatterGalaxyModal, PriorityPresets.confirm(1));
+Modal.bigCrunch = new Modal(BigCrunchModal, PriorityPresets.confirm(1));
+Modal.replicantiGalaxy = new Modal(ReplicantiGalaxyModal, PriorityPresets.confirm(1));
+Modal.eternity = new Modal(EternityModal, PriorityPresets.confirm(1));
+Modal.enterDilation = new Modal(EnterDilationModal, PriorityPresets.confirm(1));
+Modal.reality = new Modal(RealityModal, PriorityPresets.confirm(1));
+Modal.resetReality = new Modal(ResetRealityModal, PriorityPresets.confirm(1));
+Modal.exitCelestialReality = new Modal(ExitCelestialModal, PriorityPresets.confirm(1));
+Modal.celestials = new Modal(EnterCelestialsModal, PriorityPresets.confirm(1));
+Modal.hardReset = new Modal(HardResetModal, PriorityPresets.confirm(1));
 Modal.enterSpeedrun = new Modal(SpeedrunModeModal);
 Modal.changeName = new Modal(ChangeNameModal);
-Modal.armageddon = new Modal(ArmageddonModal, 1);
+Modal.armageddon = new Modal(ArmageddonModal, PriorityPresets.confirm(1));
 
 Modal.confirmationOptions = new Modal(ConfirmationOptionsModal);
 Modal.infoDisplayOptions = new Modal(InfoDisplayOptionsModal);
@@ -173,16 +176,16 @@ Modal.animationOptions = new Modal(AnimationOptionsModal);
 Modal.hiddenTabs = new Modal(HiddenTabsModal);
 Modal.preferredTree = new Modal(PreferredTreeModal);
 
-Modal.deleteCompanion = new Modal(DeleteCompanionGlyphModal, 1);
-Modal.glyphDelete = new Modal(DeleteGlyphModal, 1);
-Modal.glyphPurge = new Modal(PurgeGlyphModal, 1);
-Modal.glyphSacrifice = new Modal(SacrificeGlyphModal, 1);
-Modal.glyphRefine = new Modal(RefineGlyphModal, 1);
-Modal.deleteAllUnprotectedGlyphs = new Modal(PurgeAllUnprotectedGlyphsModal, 1);
-Modal.deleteAllRejectedGlyphs = new Modal(PurgeAllRejectedGlyphsModal, 1);
+Modal.deleteCompanion = new Modal(DeleteCompanionGlyphModal, PriorityPresets.confirm(1));
+Modal.glyphDelete = new Modal(DeleteGlyphModal, PriorityPresets.confirm(1));
+Modal.glyphPurge = new Modal(PurgeGlyphModal, PriorityPresets.confirm(1));
+Modal.glyphSacrifice = new Modal(SacrificeGlyphModal, PriorityPresets.confirm(1));
+Modal.glyphRefine = new Modal(RefineGlyphModal, PriorityPresets.confirm(1));
+Modal.deleteAllUnprotectedGlyphs = new Modal(PurgeAllUnprotectedGlyphsModal, PriorityPresets.confirm(1));
+Modal.deleteAllRejectedGlyphs = new Modal(PurgeAllRejectedGlyphsModal, PriorityPresets.confirm(1));
 
 Modal.glyphShowcasePanel = new Modal(GlyphShowcasePanelModal);
-Modal.glyphUndo = new Modal(UndoGlyphModal, 1);
+Modal.glyphUndo = new Modal(UndoGlyphModal, PriorityPresets.confirm(1));
 Modal.glyphReplace = new Modal(ReplaceGlyphModal);
 Modal.enslavedHints = new Modal(EnslavedHintsModal);
 Modal.realityGlyph = new Modal(RealityGlyphCreationModal);
@@ -200,8 +203,8 @@ Modal.shop = new Modal(StdStoreModal);
 Modal.studyString = new Modal(StudyStringModal);
 Modal.singularityMilestones = new Modal(SingularityMilestonesModal);
 Modal.pelleEffects = new Modal(PelleEffectsModal);
-Modal.sacrifice = new Modal(SacrificeModal, 1);
-Modal.breakInfinity = new Modal(BreakInfinityModal);
+Modal.sacrifice = new Modal(SacrificeModal, PriorityPresets.confirm(1));
+Modal.breakInfinity = new Modal(BreakInfinityModal, PriorityPresets.confirm(1));
 Modal.celestialQuote = new class extends Modal {
   show(celestial, lines) {
     if (!GameUI.initialized || player.speedrun.isActive) return;
@@ -250,7 +253,7 @@ Modal.celestialQuote = new class extends Modal {
     }
     return x;
   }
-}(CelestialQuoteModal, 2, true);
+}(CelestialQuoteModal, { priority: 2 }, true);
 
 Modal.cloudSaveConflict = new Modal(CloudSaveConflictModal);
 Modal.cloudLoadConflict = new Modal(CloudLoadConflictModal);
@@ -324,4 +327,4 @@ Modal.message = new class extends Modal {
       this.closeButton = this.queue[0].closeButton;
     }
   }
-}(MessageModal, 2);
+}(MessageModal, { priority: 2 });
